[refactor] Default Table Container 컴포넌트 디렉토리 변경
Desc
- TableProvider에 기본 값으로 적용하는 Table Container 컴포넌트 디렉토리 이동 (provider -> component)
- 분류 상 더 적합하다고 판단하여 수정
